refactor: move identify error component
