+++ application/ui/src/features/robots/robot-form/robot-schema/components/connection-field.tsx
@@ -1,11 +1,10 @@
 import { ActionButton, ComboBox, Flex, Icon, Item, Text, View } from '@geti-ui/ui';
 import { Refresh } from '@geti-ui/ui/icons';
 
-import { getApiErrorMessage, isSerialPermissionDeniedError } from '../../../../../api/errors';
 import { useCatalogIdentifyMutation, useDiscoverRobotsQuery } from '../../../robot-catalog.hooks';
 import { SchemaRobotType } from '../../../robot-types';
-import { InlineAlert } from '../../../setup-wizard/shared/inline-alert';
 import { ConnectionItem } from '../types';
+import { IdentifyError } from './identify-error';
 
 type Device = { serial_number: string | null; connection_string: string | null };
 
@@ -62,25 +61,6 @@ type ConnectionFieldProps = {
     payload: Record<string, unknown>;
     options: ConnectionItem;
     onChange: (field: string, value: unknown) => void;
-};
-
-export const IdentifyError = ({ error }: { error: unknown }) => {
-    if (isSerialPermissionDeniedError(error)) {
-        return (
-            <InlineAlert variant='error'>
-                <strong>Permission Denied</strong>: The application does not have permission to access the robot&apos;s
-                USB port.
-            </InlineAlert>
-        );
-    }
-
-    return (
-        <InlineAlert variant='error'>
-            <strong>Identify Failed</strong>:{' '}
-            {getApiErrorMessage(error) ??
-                'The robot could not be identified. Make sure it is powered on and not already in use, then try again.'}
-        </InlineAlert>
-    );
 };
 
 export const ConnectionField = ({ robotType, payload, options, onChange }: ConnectionFieldProps) => {
